Move basic json types into @iov/encoding
