+++ packages/iov-jsonrpc/types/index.d.ts
@@ -1,5 +1,4 @@
 export { makeJsonRpcId } from "./id";
 export { JsonRpcClient, SimpleMessagingConnection } from "./jsonrpcclient";
-export { JsonCompatibleArray, JsonCompatibleDictionary, JsonCompatibleValue, isJsonCompatibleArray, isJsonCompatibleDictionary, isJsonCompatibleValue, } from "./jsoncompatibledictionary";
 export { parseJsonRpcId, parseJsonRpcRequest, parseJsonRpcResponse2, parseJsonRpcErrorResponse, parseJsonRpcSuccessResponse, } from "./parse";
 export { isJsonRpcErrorResponse, isJsonRpcSuccessResponse, JsonRpcError, JsonRpcErrorResponse, JsonRpcId, JsonRpcRequest, JsonRpcResponse, JsonRpcSuccessResponse, jsonRpcCode, jsonRpcCodeInternalError, jsonRpcCodeInvalidParams, jsonRpcCodeInvalidRequest, jsonRpcCodeMethodNotFound, jsonRpcCodeParseError, jsonRpcCodeServerErrorDefault, } from "./types";
